requirements.txt
discord.py
requests
mailslurp_client
jishaku
fake-useragent